requirements.txt
Flask==3.0.0
flask_login==0.6.3
flask_mail==0.9.1
flask_sqlalchemy==3.1.1
numpy==1.26.3
pandas==2.1.4
python-dotenv==1.0.0
scikit_learn==1.3.2
Werkzeug==3.0.1
uvicorn==0.15.0
